No.punq/requirements.txt
discord.py>=2.4.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiofiles>=23.2.0
jinja2>=3.1.0
aiohttp==3.9.5
itsdangerous>=2.1.0